Move JWT from apiserver package to common
